\name{isolate}
\alias{isolate}
\docType{data}
\title{
Example mass spectra metadata
}
\description{
Vector containing mass spectra isolate IDs.
}
\usage{data(isolate)}
\format{
  The format is:
 Factor w/ 9 levels "33","43","48",..: 1 1 1 1 1 1 1 1 1 1 ...
}
\details{
Note that these data are simply a random subset of a larger real data set. They are only provided for illustration purposes and results from using them cannot be interpreted in any sensible way.
}
\examples{
data(isolate)
str(isolate)
}
\keyword{datasets}
