\name{MALDIrppa-package}
\alias{MALDIrppa-package}
\alias{MALDIrppa}
\docType{package}
\title{
MALDI mass spectra robust pre-processing and analysis
}
\description{
This package is aimed to provide functions for robust pre-processing and analysis of MALDI mass spectrometry data and other data management tools based on objects and procedures from \code{\link{MALDIquant}}.
}
\details{
\tabular{ll}{
Package: \tab MALDIrppa\cr
Type: \tab Package\cr
Version: \tab 0.5.0\cr
Date: \tab 2016-10\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Javier Palarea-Albaladejo

Maintainer: Javier Palarea-Albaladejo <javier.palarea@bioss.ac.uk>
}
\references{

}
\keyword{mass spectrometry, pre-processing, robust statistics}
\seealso{

}
\examples{
}
