\name{spectra}
\alias{spectra}
\docType{data}
\title{
Example mass spectra profiles
}
\description{
List of mass spectra (\code{\link{MassSpectrum}} class).  
}
\usage{data(spectra)}

\details{
Low-resolution mass spectrometry data set in the range [2500, 13000] m/z provided for illustration purposes only.
}

\examples{
data(spectra)

str(spectra[[1]])
# plot(spectra[[1]])
}
\keyword{datasets}
