\name{type}
\alias{type}
\docType{data}
\title{
Example mass spectra metadata
}
\description{
Vector containing mass spectra IDs.
}
\usage{data(type)}
\format{
  The format is:
 Factor w/ 350 levels "230212A0101",..: 1 2 3 4 5 6 7 8 9 10 ...
}
\details{
Note that these data are simply a random subset of a larger real data set. They are only provided for illustration purposes and results from using them cannot be interpreted in any sensible way.
}

\examples{
data(type)
str(type)
}
\keyword{datasets}
