\name{MALDIrppa-package}
\alias{MALDIrppa-package}
\alias{MALDIrppa}
\docType{package}
\title{
MALDI mass spec robust pre-processing and analysis
}
\description{
Functions for robust MALDI mass spectrometry data pre-processing and analysis based on the MALDIquant R package and other helper functions.
}
\details{
\tabular{ll}{
Package: \tab MALDIrppa\cr
Type: \tab Package\cr
Version: \tab 0.4.0\cr
Date: \tab 2016-02\cr
License: \tab GPL (>= 2)\cr
}
This package provides functions for MALDI mass spectrometry data robust pre-processing and analysis based on the MALDIquant R package and other helper functions.
}
\author{
Javier Palarea-Albaladejo

Maintainer: Javier Palarea-Albaladejo <javier.palarea@bioss.ac.uk>
}
\references{

}
\keyword{mass spectrometry, pre-processing, robust}
\seealso{

}
\examples{
}
