\name{MALDIrppa-package}
\alias{MALDIrppa-package}
\alias{MALDIrppa}
\docType{package}
\title{
MALDI mass spectra robust pre-processing and analysis
}
\description{
This package provides procedures for quality control and robust pre-processing and analysis of MALDI mass spectrometry data based on objects and methods from the \code{\link{MALDIquant}} package. Moreover, it includes some additional functionalities and data summary and management tools (see vignette).
}
\details{
\tabular{ll}{
Package: \tab MALDIrppa\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2022-02\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Javier Palarea-Albaladejo

Maintainer: Javier Palarea-Albaladejo <javier.palarea@udg.edu>
}
