\name{writeIntensity}
\alias{writeIntensity}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Write intensity matrix in different formats
}
\description{
This function writes an intensity matrix as generated by \code{\link{intensityMatrix}} into a file in the R, csv or NEXUS format. For NEXUS format it allows to specify sample (taxa) labels and peak weights for subsequent analysis.
}
\usage{
writeIntensity(x, filename = "intMatrix", format = c("R", "csv", "NEXUS"),
               binary = FALSE, labels = NULL, weights=NULL, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Intensity matrix as obtained from \code{\link{intensityMatrix}}.
}
  \item{filename}{
A character string specifying a name for the destination file (filename extension not required).
}
  \item{format}{
One of R (default \code{.Rdata} file), text (comma-separated \code{.csv} file) or NEXUS (\code{.nex} file).
}
  \item{binary}{
Logical value. If \code{TRUE}, a binary version (1: peak presence, 0: peak absence) of \code{x} is saved (default \code{FALSE}).
}
  \item{labels}{
Optional vector of names for the samples.
}
  \item{weights}{
Optional numeric vector of peak weights.
}
  \item{\dots}{
Additional arguments for \code{\link{save}} and \code{\link{write.table}}. For NEXUS format options see \code{write.nexus.data} in the \code{ape} library (e.g. choices for \code{datablock} or \code{interleaved}).
}
}
\details{
This is a wrapper function to simplify the writing of an intensity matrix in different formats while adding some extra features. It includes the NEXUS format as an extension of \code{write.nexus.data} in the \code{ape} package to handle peak intensity data. It also allows for taxa/samples names and pre-computed peak weights to be included in the NEXUS file. It checks whether the names meet NEXUS name conventions and gives them adequate format if not. A binary intensity matrix is always internally generated (\code{binary = TRUE}) when the NEXUS format is chosen. If any, \code{NA} values in \code{x} are assumed to denote zero intensity/peak absence and are then internally converted into zeros.
}

\examples{
# Load example data

data(spectra) # list of MassSpectra class objects

# Some pre-processing

spectra <- screenSpectra(spectra)$fspectra
spectra <- transformIntensity(spectra, method = "sqrt")
spectra <- wavSmoothing(spectra)
spectra <- removeBaseline(spectra)
peaks <- detectPeaks(spectra)
peaks <- alignPeaks(peaks, minFreq = 0.5)

# Intensity matrix

int <- intensityMatrix(peaks)

# default R format

writeIntensity(int,file="int")

# binary NEXUS file

writeIntensity(int,file="int.binary",format="NEXUS",interleaved=FALSE)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
